src/Olbrasoft.Blog.Data.EntityFrameworkCore/DesignTimeBlogDbContextFactory.cs
using Microsoft.EntityFrameworkCore;
using Microsoft.EntityFrameworkCore.Design;
using Microsoft.Extensions.Configuration;
using System;
using System.IO;

namespace Olbrasoft.Blog.Data.EntityFrameworkCore
{
    public class DesignTimeBlogDbContextFactory : IDesignTimeDbContextFactory<BlogDbContext>
    {
        public BlogDbContext CreateDbContext(string[] args)
        {
            // Get environment
            var environment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");

            // Build config
            IConfiguration config = new ConfigurationBuilder()
                .SetBasePath(Path.Combine(Directory.GetCurrentDirectory()))
                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
                .AddJsonFile($"appsettings.{environment}.json", optional: true)
                .AddEnvironmentVariables()
                .Build();

            var connectionString = config.GetConnectionString("BlogDbConnectionString");

            // Create DB context with connection from your AppSettings
            var optionsBuilder = new DbContextOptionsBuilder<BlogDbContext>()
                .UseSqlServer(connectionString);

            return new BlogDbContext(optionsBuilder.Options);
        }
    }
}